+++ hikari/audit_logs.py
@@ -47,7 +47,7 @@ from hikari import snowflakes
 from hikari.internal import attrs_extensions
 from hikari.internal import collections
 from hikari.internal import enums
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     import datetime
