+++ hikari/channels.py
@@ -77,7 +77,7 @@ from hikari import webhooks
 from hikari.internal import attrs_extensions
 from hikari.internal import enums
 from hikari.internal import routes
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     import datetime
