+++ hikari/emojis.py
@@ -35,7 +35,7 @@ from hikari import snowflakes
 from hikari import urls
 from hikari.internal import attrs_extensions
 from hikari.internal import routes
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     from hikari import traits
