+++ hikari/events/base_events.py
@@ -40,7 +40,7 @@ import attrs
 
 from hikari.api import shard as gateway_shard
 from hikari.internal import attrs_extensions
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     import types
