+++ hikari/events/message_events.py
@@ -49,7 +49,7 @@ from hikari import undefined
 from hikari.events import base_events
 from hikari.events import shard_events
 from hikari.internal import attrs_extensions
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     from hikari import embeds as embeds_
