+++ hikari/events/shard_events.py
@@ -41,7 +41,7 @@ import attrs
 from hikari.events import base_events
 from hikari.internal import attrs_extensions
 from hikari.internal import collections
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     from hikari import applications
