+++ hikari/events/user_events.py
@@ -30,7 +30,7 @@ import attrs
 
 from hikari.events import shard_events
 from hikari.internal import attrs_extensions
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     from hikari import traits
