+++ hikari/impl/buckets.py
@@ -202,7 +202,7 @@ from hikari.impl import rate_limits
 from hikari.internal import routes
 from hikari.internal import time
 from hikari.internal import ux
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     import types
