+++ hikari/impl/config.py
@@ -40,7 +40,7 @@ import attrs
 from hikari.api import config
 from hikari.internal import attrs_extensions
 from hikari.internal import data_binding
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 _BASICAUTH_TOKEN_PREFIX: typing.Final[str] = "Basic"  # noqa: S105
 _PROXY_AUTHENTICATION_HEADER: typing.Final[str] = "Proxy-Authentication"
