+++ hikari/impl/entity_factory.py
@@ -67,7 +67,7 @@ from hikari.interactions import modal_interactions
 from hikari.internal import attrs_extensions
 from hikari.internal import data_binding
 from hikari.internal import time
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     ValueT = typing.TypeVar("ValueT")
