+++ hikari/impl/event_manager_base.py
@@ -46,7 +46,7 @@ from hikari.internal import aio
 from hikari.internal import fast_protocol
 from hikari.internal import reflect
 from hikari.internal import ux
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     from typing_extensions import Self
