+++ hikari/impl/gateway_bot.py
@@ -52,7 +52,7 @@ from hikari.internal import data_binding
 from hikari.internal import signals
 from hikari.internal import time
 from hikari.internal import ux
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     import concurrent.futures
