+++ hikari/impl/rate_limits.py
@@ -41,7 +41,7 @@ import random
 import typing
 
 from hikari.internal import time
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     import types
