+++ hikari/impl/shard.py
@@ -49,7 +49,7 @@ from hikari.internal import data_binding
 from hikari.internal import net
 from hikari.internal import time
 from hikari.internal import ux
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     import datetime
