+++ hikari/impl/special_endpoints.py
@@ -74,7 +74,7 @@ from hikari.internal import data_binding
 from hikari.internal import mentions
 from hikari.internal import routes
 from hikari.internal import time
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     import concurrent.futures
