+++ hikari/impl/voice.py
@@ -34,7 +34,7 @@ from hikari import snowflakes
 from hikari.api import voice
 from hikari.events import voice_events
 from hikari.internal import ux
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     from hikari import channels
