+++ hikari/internal/cache.py
@@ -62,7 +62,7 @@ from hikari import voices
 from hikari.api import cache
 from hikari.internal import attrs_extensions
 from hikari.internal import collections
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     import datetime
