+++ hikari/internal/collections.py
@@ -40,7 +40,7 @@ import sys
 import typing
 
 from hikari import snowflakes
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     from typing_extensions import Self
