+++ hikari/internal/enums.pyi
@@ -36,7 +36,7 @@ from typing import TypeVar as __TypeVar
 
 from typing_extensions import Self as __Self
 
-from hikari.internal.override import override as __override
+from hikari.internal.typing_backport import override as __override
 
 Enum = __enum.Enum
