+++ hikari/internal/ux.py
@@ -48,7 +48,7 @@ import colorlog.escape_codes
 from hikari import _about as about
 from hikari.internal import data_binding
 from hikari.internal import net
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     from hikari.impl import config
