+++ hikari/iterators.py
@@ -34,7 +34,7 @@ import asyncio
 import typing
 
 from hikari.internal import spel
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     from _typeshed import SupportsRichComparison
