+++ hikari/snowflakes.py
@@ -38,7 +38,7 @@ import abc
 import typing
 
 from hikari.internal import time
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     import datetime
