+++ hikari/templates.py
@@ -31,7 +31,7 @@ import attrs
 from hikari import guilds
 from hikari import undefined
 from hikari.internal import attrs_extensions
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     import datetime
