+++ hikari/undefined.py
@@ -34,7 +34,7 @@ __all__: typing.Sequence[str] = (
 
 import typing
 
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     from typing_extensions import Self
