+++ hikari/users.py
@@ -36,7 +36,7 @@ from hikari import urls
 from hikari.internal import attrs_extensions
 from hikari.internal import enums
 from hikari.internal import routes
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if not typing.TYPE_CHECKING:
     # This is insanely hacky, but it is needed for ruff to not complain until it gets type inference
