+++ hikari/voices.py
@@ -29,7 +29,7 @@ import typing
 import attrs
 
 from hikari.internal import attrs_extensions
-from hikari.internal.override import override
+from hikari.internal.typing_backport import override
 
 if typing.TYPE_CHECKING:
     import datetime
